+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 lib_sw_dac change log
 =====================
 
+UNRELEASED
+----------
+
+  * FIXED: Now sends PWM zero when not fed with samples
+
 0.1.0
 -----